add new binary module (splitting out the aeron functionality
