+++ nd4j-serde/pom.xml
@@ -18,6 +18,7 @@
         <module>nd4j-camel-routes</module>
         <module>nd4j-gson</module>
         <module>nd4j-jackson-reflectionloader</module>
+        <module>nd4j-binary</module>
     </modules>
 
 </project>
